LayoutTests/http/tests/inspector/filesystem/request-filesystem-root-expected.txt

Tests requestFileSystemRoot command.

errorCode: 0
backendRootEntry.url: filesystem:http://127.0.0.1:8000/temporary/

errorCode: 8
backendRootEntry: (null)

errorCode: 2
backendRootEntry: (null)

All tests have done.
